Filter docs for Table
