--- components/doc/datatable/filter/filtermenudoc.js
@@ -1,797 +1,0 @@
-import React, { useEffect, useState } from 'react';
-import { CustomerService } from '../../../../service/CustomerService';
-import { FilterMatchMode, FilterOperator } from '../../../lib/api/Api';
-import { Button } from '../../../lib/button/Button';
-import { Calendar } from '../../../lib/calendar/Calendar';
-import { Column } from '../../../lib/column/Column';
-import { DataTable } from '../../../lib/datatable/DataTable';
-import { Dropdown } from '../../../lib/dropdown/Dropdown';
-import { InputNumber } from '../../../lib/inputnumber/InputNumber';
-import { InputText } from '../../../lib/inputtext/InputText';
-import { MultiSelect } from '../../../lib/multiselect/MultiSelect';
-import { ProgressBar } from '../../../lib/progressbar/ProgressBar';
-import { Slider } from '../../../lib/slider/Slider';
-import { TriStateCheckbox } from '../../../lib/tristatecheckbox/TriStateCheckbox';
-import { classNames } from '../../../lib/utils/Utils';
-import { DocSectionCode } from '../../common/docsectioncode';
-import { DocSectionText } from '../../common/docsectiontext';
-
-export function FilterMenuDoc(props) {
-    const [customers, setCustomers] = useState(null);
-    const [filters, setFilters] = useState(null);
-    const [globalFilterValue, setGlobalFilterValue] = useState('');
-    const [loading, setLoading] = useState(true);
-    const representatives = [
-        { name: 'Amy Elsner', image: 'amyelsner.png' },
-        { name: 'Anna Fali', image: 'annafali.png' },
-        { name: 'Asiya Javayant', image: 'asiyajavayant.png' },
-        { name: 'Bernardo Dominic', image: 'bernardodominic.png' },
-        { name: 'Elwin Sharvill', image: 'elwinsharvill.png' },
-        { name: 'Ioni Bowcher', image: 'ionibowcher.png' },
-        { name: 'Ivan Magalhaes', image: 'ivanmagalhaes.png' },
-        { name: 'Onyama Limba', image: 'onyamalimba.png' },
-        { name: 'Stephen Shaw', image: 'stephenshaw.png' },
-        { name: 'XuXue Feng', image: 'xuxuefeng.png' }
-    ];
-
-    const statuses = ['unqualified', 'qualified', 'new', 'negotiation', 'renewal', 'proposal'];
-
-    useEffect(() => {
-        CustomerService.getCustomersLarge().then((data) => {
-            setCustomers(getCustomers(data));
-            setLoading(false);
-        });
-        initFilters();
-    }, []); // eslint-disable-line react-hooks/exhaustive-deps
-
-    const getCustomers = (data) => {
-        return [...(data || [])].map((d) => {
-            d.date = new Date(d.date);
-
-            return d;
-        });
-    };
-
-    const formatDate = (value) => {
-        return value.toLocaleDateString('en-US', {
-            day: '2-digit',
-            month: '2-digit',
-            year: 'numeric'
-        });
-    };
-
-    const formatCurrency = (value) => {
-        return value.toLocaleString('en-US', { style: 'currency', currency: 'USD' });
-    };
-
-    const clearFilter = () => {
-        initFilters();
-    };
-
-    const onGlobalFilterChange = (e) => {
-        const value = e.target.value;
-        let _filters = { ...filters };
-
-        _filters['global'].value = value;
-
-        setFilters(_filters);
-        setGlobalFilterValue(value);
-    };
-
-    const initFilters = () => {
-        setFilters({
-            global: { value: null, matchMode: FilterMatchMode.CONTAINS },
-            name: { operator: FilterOperator.AND, constraints: [{ value: null, matchMode: FilterMatchMode.STARTS_WITH }] },
-            'country.name': { operator: FilterOperator.AND, constraints: [{ value: null, matchMode: FilterMatchMode.STARTS_WITH }] },
-            representative: { value: null, matchMode: FilterMatchMode.IN },
-            date: { operator: FilterOperator.AND, constraints: [{ value: null, matchMode: FilterMatchMode.DATE_IS }] },
-            balance: { operator: FilterOperator.AND, constraints: [{ value: null, matchMode: FilterMatchMode.EQUALS }] },
-            status: { operator: FilterOperator.OR, constraints: [{ value: null, matchMode: FilterMatchMode.EQUALS }] },
-            activity: { value: null, matchMode: FilterMatchMode.BETWEEN },
-            verified: { value: null, matchMode: FilterMatchMode.EQUALS }
-        });
-        setGlobalFilterValue('');
-    };
-
-    const renderHeader = () => {
-        return (
-            <div className="flex justify-content-between">
-                <Button type="button" icon="pi pi-filter-slash" label="Clear" className="p-button-outlined" onClick={clearFilter} />
-                <span className="p-input-icon-left">
-                    <i className="pi pi-search" />
-                    <InputText value={globalFilterValue} onChange={onGlobalFilterChange} placeholder="Keyword Search" />
-                </span>
-            </div>
-        );
-    };
-
-    const countryBodyTemplate = (rowData) => {
-        return (
-            <React.Fragment>
-                <img alt="flag" src="https://primefaces.org/cdn/primereact/images/flag/flag_placeholder.png" className={`flag flag-${rowData.country.code}`} width={30} />
-                <span className="vertical-align-middle ml-2">{rowData.country.name}</span>
-            </React.Fragment>
-        );
-    };
-
-    const filterClearTemplate = (options) => {
-        return <Button type="button" icon="pi pi-times" onClick={options.filterClearCallback} className="p-button-secondary"></Button>;
-    };
-
-    const filterApplyTemplate = (options) => {
-        return <Button type="button" icon="pi pi-check" onClick={options.filterApplyCallback} className="p-button-success"></Button>;
-    };
-
-    const filterFooterTemplate = () => {
-        return <div className="px-3 pt-0 pb-3 text-center font-bold">Customized Buttons</div>;
-    };
-
-    const representativeBodyTemplate = (rowData) => {
-        const representative = rowData.representative;
-
-        return (
-            <React.Fragment>
-                <img alt={representative.name} src={`https://primefaces.org/cdn/primereact/images/avatar/${representative.image}`} width={32} style={{ verticalAlign: 'middle' }} />
-                <span className="vertical-align-middle ml-2">{representative.name}</span>
-            </React.Fragment>
-        );
-    };
-
-    const representativeFilterTemplate = (options) => {
-        return <MultiSelect value={options.value} options={representatives} itemTemplate={representativesItemTemplate} onChange={(e) => options.filterCallback(e.value)} optionLabel="name" placeholder="Any" className="p-column-filter" />;
-    };
-
-    const representativesItemTemplate = (option) => {
-        return (
-            <div className="p-multiselect-representative-option">
-                <img alt={option.name} src={`https://primefaces.org/cdn/primereact/images/avatar/${option.image}`} width={32} style={{ verticalAlign: 'middle' }} />
-                <span className="vertical-align-middle ml-2">{option.name}</span>
-            </div>
-        );
-    };
-
-    const dateBodyTemplate = (rowData) => {
-        return formatDate(rowData.date);
-    };
-
-    const dateFilterTemplate = (options) => {
-        return <Calendar value={options.value} onChange={(e) => options.filterCallback(e.value, options.index)} dateFormat="mm/dd/yy" placeholder="mm/dd/yyyy" mask="99/99/9999" />;
-    };
-
-    const balanceBodyTemplate = (rowData) => {
-        return formatCurrency(rowData.balance);
-    };
-
-    const balanceFilterTemplate = (options) => {
-        return <InputNumber value={options.value} onChange={(e) => options.filterCallback(e.value, options.index)} mode="currency" currency="USD" locale="en-US" />;
-    };
-
-    const statusBodyTemplate = (rowData) => {
-        return <span className={`customer-badge status-${rowData.status}`}>{rowData.status}</span>;
-    };
-
-    const statusFilterTemplate = (options) => {
-        return <Dropdown value={options.value} options={statuses} onChange={(e) => options.filterCallback(e.value, options.index)} itemTemplate={statusItemTemplate} placeholder="Select a Status" className="p-column-filter" showClear />;
-    };
-
-    const statusItemTemplate = (option) => {
-        return <span className={`customer-badge status-${option}`}>{option}</span>;
-    };
-
-    const activityBodyTemplate = (rowData) => {
-        return <ProgressBar value={rowData.activity} showValue={false}></ProgressBar>;
-    };
-
-    const activityFilterTemplate = (options) => {
-        return (
-            <React.Fragment>
-                <Slider value={options.value} onChange={(e) => options.filterCallback(e.value)} range className="m-3"></Slider>
-                <div className="flex align-items-center justify-content-between px-2">
-                    <span>{options.value ? options.value[0] : 0}</span>
-                    <span>{options.value ? options.value[1] : 100}</span>
-                </div>
-            </React.Fragment>
-        );
-    };
-
-    const verifiedBodyTemplate = (rowData) => {
-        return <i className={classNames('pi', { 'true-icon pi-check-circle': rowData.verified, 'false-icon pi-times-circle': !rowData.verified })}></i>;
-    };
-
-    const verifiedFilterTemplate = (options) => {
-        return <TriStateCheckbox value={options.value} onChange={(e) => options.filterCallback(e.value)} />;
-    };
-
-    const header = renderHeader();
-
-    const code = {
-        basic: `
-<DataTable value={customers} paginator className="p-datatable-customers" showGridlines rows={10}
-    dataKey="id" filters={filters} filterDisplay="menu" loading={loading} responsiveLayout="scroll"
-    globalFilterFields={['name', 'country.name', 'representative.name', 'balance', 'status']} header={header} emptyMessage="No customers found.">
-    <Column field="name" header="Name" filter filterPlaceholder="Search by name" style={{ minWidth: '12rem' }} />
-    <Column header="Country" filterField="country.name" style={{ minWidth: '12rem' }} body={countryBodyTemplate} filter filterPlaceholder="Search by country"
-        filterClear={filterClearTemplate} filterApply={filterApplyTemplate} filterFooter={filterFooterTemplate} />
-    <Column header="Agent" filterField="representative" showFilterMatchModes={false} filterMenuStyle={{ width: '14rem'}} style={{ minWidth: '14rem' }} body={representativeBodyTemplate}
-        filter filterElement={representativeFilterTemplate} />
-    <Column header="Date" filterField="date" dataType="date" style={{ minWidth: '10rem' }} body={dateBodyTemplate}
-        filter filterElement={dateFilterTemplate} />
-    <Column header="Balance" filterField="balance" dataType="numeric" style={{ minWidth: '10rem' }} body={balanceBodyTemplate} filter filterElement={balanceFilterTemplate} />
-    <Column field="status" header="Status" filterMenuStyle={{ width: '14rem' }} style={{ minWidth: '12rem' }} body={statusBodyTemplate} filter filterElement={statusFilterTemplate} />
-    <Column field="activity" header="Activity" showFilterMatchModes={false} style={{ minWidth: '12rem' }} body={activityBodyTemplate} filter filterElement={activityFilterTemplate} />
-    <Column field="verified" header="Verified" dataType="boolean" bodyClassName="text-center" style={{ minWidth: '8rem' }} body={verifiedBodyTemplate} filter filterElement={verifiedFilterTemplate} />
-</DataTable>
-        `,
-        javascript: `
-import React, { useState, useEffect } from 'react';
-import { classNames } from 'primereact/utils';
-import { FilterMatchMode, FilterOperator } from 'primereact/api';
-import { DataTable } from 'primereact/datatable';
-import { Column } from 'primereact/column';
-import { InputText } from 'primereact/inputtext';
-import { Dropdown } from 'primereact/dropdown';
-import { InputNumber } from 'primereact/inputnumber';
-import { Button } from 'primereact/button';
-import { ProgressBar } from 'primereact/progressbar';
-import { Calendar } from 'primereact/calendar';
-import { MultiSelect } from 'primereact/multiselect';
-import { Slider } from 'primereact/slider';
-import { TriStateCheckbox } from 'primereact/tristatecheckbox';
-import { CustomerService } from './service/CustomerService';
-import './DataTableDemo.css';
-
-export default function FilterMenuDoc() {
-    const [customers, setCustomers] = useState(null);
-    const [filters, setFilters] = useState(null);
-    const [globalFilterValue, setGlobalFilterValue] = useState('');
-    const [loading, setLoading] = useState(true);
-    const representatives = [
-        {name: "Amy Elsner", image: 'amyelsner.png'},
-        {name: "Anna Fali", image: 'annafali.png'},
-        {name: "Asiya Javayant", image: 'asiyajavayant.png'},
-        {name: "Bernardo Dominic", image: 'bernardodominic.png'},
-        {name: "Elwin Sharvill", image: 'elwinsharvill.png'},
-        {name: "Ioni Bowcher", image: 'ionibowcher.png'},
-        {name: "Ivan Magalhaes",image: 'ivanmagalhaes.png'},
-        {name: "Onyama Limba", image: 'onyamalimba.png'},
-        {name: "Stephen Shaw", image: 'stephenshaw.png'},
-        {name: "XuXue Feng", image: 'xuxuefeng.png'}
-    ];
-
-    const statuses = [
-        'unqualified', 'qualified', 'new', 'negotiation', 'renewal', 'proposal'
-    ];
-
-    
-
-    useEffect(() => {
-        CustomerService.getCustomersLarge().then(data => { setCustomers(getCustomers(data)); setLoading(false) });
-        initFilters();
-    }, []); // eslint-disable-line react-hooks/exhaustive-deps
-
-
-    const getCustomers = (data) => {
-        return [...data || []].map(d => {
-            d.date = new Date(d.date);
-            return d;
-        });
-    }
-
-    const formatDate = (value) => {
-        return value.toLocaleDateString('en-US', {
-            day: '2-digit',
-            month: '2-digit',
-            year: 'numeric',
-        });
-    }
-
-    const formatCurrency = (value) => {
-        return value.toLocaleString('en-US', { style: 'currency', currency: 'USD' });
-    }
-
-    const clearFilter = () => {
-        initFilters();
-    }
-
-    const onGlobalFilterChange = (e) => {
-        const value = e.target.value;
-        let _filters = { ...filters };
-        _filters['global'].value = value;
-
-        setFilters(_filters);
-        setGlobalFilterValue(value);
-    }
-
-    const initFilters = () => {
-        setFilters({
-            'global': { value: null, matchMode: FilterMatchMode.CONTAINS },
-            'name': { operator: FilterOperator.AND, constraints: [{ value: null, matchMode: FilterMatchMode.STARTS_WITH }] },
-            'country.name': { operator: FilterOperator.AND, constraints: [{ value: null, matchMode: FilterMatchMode.STARTS_WITH }] },
-            'representative': { value: null, matchMode: FilterMatchMode.IN },
-            'date': { operator: FilterOperator.AND, constraints: [{ value: null, matchMode: FilterMatchMode.DATE_IS }] },
-            'balance': { operator: FilterOperator.AND, constraints: [{ value: null, matchMode: FilterMatchMode.EQUALS }] },
-            'status': { operator: FilterOperator.OR, constraints: [{ value: null, matchMode: FilterMatchMode.EQUALS }] },
-            'activity': { value: null, matchMode: FilterMatchMode.BETWEEN },
-            'verified': { value: null, matchMode: FilterMatchMode.EQUALS }
-        });
-        setGlobalFilterValue('');
-    }
-
-    const renderHeader = () => {
-        return (
-            <div className="flex justify-content-between">
-                <Button type="button" icon="pi pi-filter-slash" label="Clear" className="p-button-outlined" onClick={clearFilter} />
-                <span className="p-input-icon-left">
-                    <i className="pi pi-search" />
-                    <InputText value={globalFilterValue} onChange={onGlobalFilterChange} placeholder="Keyword Search" />
-                </span>
-            </div>
-        )
-    }
-
-    const countryBodyTemplate = (rowData) => {
-        return (
-            <React.Fragment>
-                <img alt="flag" src="https://primefaces.org/cdn/primereact/images/flag/flag_placeholder.png" className={\`flag flag-\${rowData.country.code}\`} width={30} />
-                <span className="vertical-align-middle ml-2">{rowData.country.name}</span>
-            </React.Fragment>
-        );
-    }
-
-    const filterClearTemplate = (options) => {
-        return <Button type="button" icon="pi pi-times" onClick={options.filterClearCallback} className="p-button-secondary"></Button>;
-    }
-
-    const filterApplyTemplate = (options) => {
-        return <Button type="button" icon="pi pi-check" onClick={options.filterApplyCallback} className="p-button-success"></Button>
-    }
-
-    const filterFooterTemplate = () => {
-        return <div className="px-3 pt-0 pb-3 text-center font-bold">Customized Buttons</div>;
-    }
-
-    const representativeBodyTemplate = (rowData) => {
-        const representative = rowData.representative;
-        return (
-            <React.Fragment>
-                <img alt={representative.name} src={\`https://primefaces.org/cdn/primereact/images/avatar/\${representative.image}\`} width={32} style={{ verticalAlign: 'middle' }} />
-                <span className="vertical-align-middle ml-2">{representative.name}</span>
-            </React.Fragment>
-        );
-    }
-
-    const representativeFilterTemplate = (options) => {
-        return <MultiSelect value={options.value} options={representatives} itemTemplate={representativesItemTemplate} onChange={(e) => options.filterCallback(e.value)} optionLabel="name" placeholder="Any" className="p-column-filter" />;
-    }
-
-    const representativesItemTemplate = (option) => {
-        return (
-            <div className="p-multiselect-representative-option">
-                <img alt={option.name} src={\`https://primefaces.org/cdn/primereact/images/avatar/\${option.image}\`} width={32} style={{ verticalAlign: 'middle' }} />
-                <span className="vertical-align-middle ml-2">{option.name}</span>
-            </div>
-        );
-    }
-
-    const dateBodyTemplate = (rowData) => {
-        return formatDate(rowData.date);
-    }
-
-    const dateFilterTemplate = (options) => {
-        return <Calendar value={options.value} onChange={(e) => options.filterCallback(e.value, options.index)} dateFormat="mm/dd/yy" placeholder="mm/dd/yyyy" mask="99/99/9999" />
-    }
-
-    const balanceBodyTemplate = (rowData) => {
-        return formatCurrency(rowData.balance);
-    }
-
-    const balanceFilterTemplate = (options) => {
-        return <InputNumber value={options.value} onChange={(e) => options.filterCallback(e.value, options.index)} mode="currency" currency="USD" locale="en-US" />
-    }
-
-    const statusBodyTemplate = (rowData) => {
-        return <span className={\`customer-badge status-\${rowData.status}\`}>{rowData.status}</span>;
-    }
-
-    const statusFilterTemplate = (options) => {
-        return <Dropdown value={options.value} options={statuses} onChange={(e) => options.filterCallback(e.value, options.index)} itemTemplate={statusItemTemplate} placeholder="Select a Status" className="p-column-filter" showClear />;
-    }
-
-    const statusItemTemplate = (option) => {
-        return <span className={\`customer-badge status-\${option}\`}>{option}</span>;
-    }
-
-    const activityBodyTemplate = (rowData) => {
-        return <ProgressBar value={rowData.activity} showValue={false}></ProgressBar>;
-    }
-
-    const activityFilterTemplate = (options) => {
-        return (
-            <React.Fragment>
-                <Slider value={options.value} onChange={(e) => options.filterCallback(e.value)} range className="m-3"></Slider>
-                <div className="flex align-items-center justify-content-between px-2">
-                    <span>{options.value ? options.value[0] : 0}</span>
-                    <span>{options.value ? options.value[1] : 100}</span>
-                </div>
-            </React.Fragment>
-        )
-    }
-
-    const verifiedBodyTemplate = (rowData) => {
-        return <i className={classNames('pi', {'true-icon pi-check-circle': rowData.verified, 'false-icon pi-times-circle': !rowData.verified})}></i>;
-    }
-
-    const verifiedFilterTemplate = (options) => {
-        return <TriStateCheckbox value={options.value} onChange={(e) => options.filterCallback(e.value)} />
-    }
-
-    const header = renderHeader();
-
-    return (
-        <div className="datatable-filter-demo">
-            <div className="card datatable-demo">
-                <DataTable value={customers} paginator className="p-datatable-customers" showGridlines rows={10}
-                    dataKey="id" filters={filters} filterDisplay="menu" loading={loading} responsiveLayout="scroll"
-                    globalFilterFields={['name', 'country.name', 'representative.name', 'balance', 'status']} header={header} emptyMessage="No customers found.">
-                    <Column field="name" header="Name" filter filterPlaceholder="Search by name" style={{ minWidth: '12rem' }} />
-                    <Column header="Country" filterField="country.name" style={{ minWidth: '12rem' }} body={countryBodyTemplate} filter filterPlaceholder="Search by country"
-                        filterClear={filterClearTemplate} filterApply={filterApplyTemplate} filterFooter={filterFooterTemplate} />
-                    <Column header="Agent" filterField="representative" showFilterMatchModes={false} filterMenuStyle={{ width: '14rem'}} style={{ minWidth: '14rem' }} body={representativeBodyTemplate}
-                        filter filterElement={representativeFilterTemplate} />
-                    <Column header="Date" filterField="date" dataType="date" style={{ minWidth: '10rem' }} body={dateBodyTemplate}
-                        filter filterElement={dateFilterTemplate} />
-                    <Column header="Balance" filterField="balance" dataType="numeric" style={{ minWidth: '10rem' }} body={balanceBodyTemplate} filter filterElement={balanceFilterTemplate} />
-                    <Column field="status" header="Status" filterMenuStyle={{ width: '14rem' }} style={{ minWidth: '12rem' }} body={statusBodyTemplate} filter filterElement={statusFilterTemplate} />
-                    <Column field="activity" header="Activity" showFilterMatchModes={false} style={{ minWidth: '12rem' }} body={activityBodyTemplate} filter filterElement={activityFilterTemplate} />
-                    <Column field="verified" header="Verified" dataType="boolean" bodyClassName="text-center" style={{ minWidth: '8rem' }} body={verifiedBodyTemplate} filter filterElement={verifiedFilterTemplate} />
-                </DataTable>
-            </div>
-        </div>
-    );
-}
-        `,
-        typescript: `
-import React, { useState, useEffect } from 'react';
-import { classNames } from 'primereact/utils';
-import { FilterMatchMode, FilterOperator } from 'primereact/api';
-import { DataTable } from 'primereact/datatable';
-import { Column } from 'primereact/column';
-import { InputText } from 'primereact/inputtext';
-import { Dropdown } from 'primereact/dropdown';
-import { InputNumber } from 'primereact/inputnumber';
-import { Button } from 'primereact/button';
-import { ProgressBar } from 'primereact/progressbar';
-import { Calendar } from 'primereact/calendar';
-import { MultiSelect } from 'primereact/multiselect';
-import { Slider } from 'primereact/slider';
-import { TriStateCheckbox } from 'primereact/tristatecheckbox';
-import { CustomerService } from './service/CustomerService';
-import './DataTableDemo.css';
-
-export default function FilterMenuDoc() {
-    const [customers, setCustomers] = useState(null);
-    const [filters, setFilters] = useState(null);
-    const [globalFilterValue, setGlobalFilterValue] = useState('');
-    const [loading, setLoading] = useState(true);
-    const representatives = [
-        {name: "Amy Elsner", image: 'amyelsner.png'},
-        {name: "Anna Fali", image: 'annafali.png'},
-        {name: "Asiya Javayant", image: 'asiyajavayant.png'},
-        {name: "Bernardo Dominic", image: 'bernardodominic.png'},
-        {name: "Elwin Sharvill", image: 'elwinsharvill.png'},
-        {name: "Ioni Bowcher", image: 'ionibowcher.png'},
-        {name: "Ivan Magalhaes",image: 'ivanmagalhaes.png'},
-        {name: "Onyama Limba", image: 'onyamalimba.png'},
-        {name: "Stephen Shaw", image: 'stephenshaw.png'},
-        {name: "XuXue Feng", image: 'xuxuefeng.png'}
-    ];
-
-    const statuses = [
-        'unqualified', 'qualified', 'new', 'negotiation', 'renewal', 'proposal'
-    ];
-
-    
-
-    useEffect(() => {
-        CustomerService.getCustomersLarge().then(data => { setCustomers(getCustomers(data)); setLoading(false) });
-        initFilters();
-    }, []); // eslint-disable-line react-hooks/exhaustive-deps
-
-
-    const getCustomers = (data) => {
-        return [...data || []].map(d => {
-            d.date = new Date(d.date);
-            return d;
-        });
-    }
-
-    const formatDate = (value) => {
-        return value.toLocaleDateString('en-US', {
-            day: '2-digit',
-            month: '2-digit',
-            year: 'numeric',
-        });
-    }
-
-    const formatCurrency = (value) => {
-        return value.toLocaleString('en-US', { style: 'currency', currency: 'USD' });
-    }
-
-    const clearFilter = () => {
-        initFilters();
-    }
-
-    const onGlobalFilterChange = (e) => {
-        const value = e.target.value;
-        let _filters = { ...filters };
-        _filters['global'].value = value;
-
-        setFilters(_filters);
-        setGlobalFilterValue(value);
-    }
-
-    const initFilters = () => {
-        setFilters({
-            'global': { value: null, matchMode: FilterMatchMode.CONTAINS },
-            'name': { operator: FilterOperator.AND, constraints: [{ value: null, matchMode: FilterMatchMode.STARTS_WITH }] },
-            'country.name': { operator: FilterOperator.AND, constraints: [{ value: null, matchMode: FilterMatchMode.STARTS_WITH }] },
-            'representative': { value: null, matchMode: FilterMatchMode.IN },
-            'date': { operator: FilterOperator.AND, constraints: [{ value: null, matchMode: FilterMatchMode.DATE_IS }] },
-            'balance': { operator: FilterOperator.AND, constraints: [{ value: null, matchMode: FilterMatchMode.EQUALS }] },
-            'status': { operator: FilterOperator.OR, constraints: [{ value: null, matchMode: FilterMatchMode.EQUALS }] },
-            'activity': { value: null, matchMode: FilterMatchMode.BETWEEN },
-            'verified': { value: null, matchMode: FilterMatchMode.EQUALS }
-        });
-        setGlobalFilterValue('');
-    }
-
-    const renderHeader = () => {
-        return (
-            <div className="flex justify-content-between">
-                <Button type="button" icon="pi pi-filter-slash" label="Clear" className="p-button-outlined" onClick={clearFilter} />
-                <span className="p-input-icon-left">
-                    <i className="pi pi-search" />
-                    <InputText value={globalFilterValue} onChange={onGlobalFilterChange} placeholder="Keyword Search" />
-                </span>
-            </div>
-        )
-    }
-
-    const countryBodyTemplate = (rowData) => {
-        return (
-            <React.Fragment>
-                <img alt="flag" src="https://primefaces.org/cdn/primereact/images/flag/flag_placeholder.png" className={\`flag flag-\${rowData.country.code}\`} width={30} />
-                <span className="vertical-align-middle ml-2">{rowData.country.name}</span>
-            </React.Fragment>
-        );
-    }
-
-    const filterClearTemplate = (options) => {
-        return <Button type="button" icon="pi pi-times" onClick={options.filterClearCallback} className="p-button-secondary"></Button>;
-    }
-
-    const filterApplyTemplate = (options) => {
-        return <Button type="button" icon="pi pi-check" onClick={options.filterApplyCallback} className="p-button-success"></Button>
-    }
-
-    const filterFooterTemplate = () => {
-        return <div className="px-3 pt-0 pb-3 text-center font-bold">Customized Buttons</div>;
-    }
-
-    const representativeBodyTemplate = (rowData) => {
-        const representative = rowData.representative;
-        return (
-            <React.Fragment>
-                <img alt={representative.name} src={\`https://primefaces.org/cdn/primereact/images/avatar/\${representative.image}\`} width={32} style={{ verticalAlign: 'middle' }} />
-                <span className="vertical-align-middle ml-2">{representative.name}</span>
-            </React.Fragment>
-        );
-    }
-
-    const representativeFilterTemplate = (options) => {
-        return <MultiSelect value={options.value} options={representatives} itemTemplate={representativesItemTemplate} onChange={(e) => options.filterCallback(e.value)} optionLabel="name" placeholder="Any" className="p-column-filter" />;
-    }
-
-    const representativesItemTemplate = (option) => {
-        return (
-            <div className="p-multiselect-representative-option">
-                <img alt={option.name} src={\`https://primefaces.org/cdn/primereact/images/avatar/\${option.image}\`} width={32} style={{ verticalAlign: 'middle' }} />
-                <span className="vertical-align-middle ml-2">{option.name}</span>
-            </div>
-        );
-    }
-
-    const dateBodyTemplate = (rowData) => {
-        return formatDate(rowData.date);
-    }
-
-    const dateFilterTemplate = (options) => {
-        return <Calendar value={options.value} onChange={(e) => options.filterCallback(e.value, options.index)} dateFormat="mm/dd/yy" placeholder="mm/dd/yyyy" mask="99/99/9999" />
-    }
-
-    const balanceBodyTemplate = (rowData) => {
-        return formatCurrency(rowData.balance);
-    }
-
-    const balanceFilterTemplate = (options) => {
-        return <InputNumber value={options.value} onChange={(e) => options.filterCallback(e.value, options.index)} mode="currency" currency="USD" locale="en-US" />
-    }
-
-    const statusBodyTemplate = (rowData) => {
-        return <span className={\`customer-badge status-\${rowData.status}\`}>{rowData.status}</span>;
-    }
-
-    const statusFilterTemplate = (options) => {
-        return <Dropdown value={options.value} options={statuses} onChange={(e) => options.filterCallback(e.value, options.index)} itemTemplate={statusItemTemplate} placeholder="Select a Status" className="p-column-filter" showClear />;
-    }
-
-    const statusItemTemplate = (option) => {
-        return <span className={\`customer-badge status-\${option}\`}>{option}</span>;
-    }
-
-    const activityBodyTemplate = (rowData) => {
-        return <ProgressBar value={rowData.activity} showValue={false}></ProgressBar>;
-    }
-
-    const activityFilterTemplate = (options) => {
-        return (
-            <React.Fragment>
-                <Slider value={options.value} onChange={(e) => options.filterCallback(e.value)} range className="m-3"></Slider>
-                <div className="flex align-items-center justify-content-between px-2">
-                    <span>{options.value ? options.value[0] : 0}</span>
-                    <span>{options.value ? options.value[1] : 100}</span>
-                </div>
-            </React.Fragment>
-        )
-    }
-
-    const verifiedBodyTemplate = (rowData) => {
-        return <i className={classNames('pi', {'true-icon pi-check-circle': rowData.verified, 'false-icon pi-times-circle': !rowData.verified})}></i>;
-    }
-
-    const verifiedFilterTemplate = (options) => {
-        return <TriStateCheckbox value={options.value} onChange={(e) => options.filterCallback(e.value)} />
-    }
-
-    const header = renderHeader();
-
-    return (
-        <div className="datatable-filter-demo">
-            <div className="card datatable-demo">
-                <DataTable value={customers} paginator className="p-datatable-customers" showGridlines rows={10}
-                    dataKey="id" filters={filters} filterDisplay="menu" loading={loading} responsiveLayout="scroll"
-                    globalFilterFields={['name', 'country.name', 'representative.name', 'balance', 'status']} header={header} emptyMessage="No customers found.">
-                    <Column field="name" header="Name" filter filterPlaceholder="Search by name" style={{ minWidth: '12rem' }} />
-                    <Column header="Country" filterField="country.name" style={{ minWidth: '12rem' }} body={countryBodyTemplate} filter filterPlaceholder="Search by country"
-                        filterClear={filterClearTemplate} filterApply={filterApplyTemplate} filterFooter={filterFooterTemplate} />
-                    <Column header="Agent" filterField="representative" showFilterMatchModes={false} filterMenuStyle={{ width: '14rem'}} style={{ minWidth: '14rem' }} body={representativeBodyTemplate}
-                        filter filterElement={representativeFilterTemplate} />
-                    <Column header="Date" filterField="date" dataType="date" style={{ minWidth: '10rem' }} body={dateBodyTemplate}
-                        filter filterElement={dateFilterTemplate} />
-                    <Column header="Balance" filterField="balance" dataType="numeric" style={{ minWidth: '10rem' }} body={balanceBodyTemplate} filter filterElement={balanceFilterTemplate} />
-                    <Column field="status" header="Status" filterMenuStyle={{ width: '14rem' }} style={{ minWidth: '12rem' }} body={statusBodyTemplate} filter filterElement={statusFilterTemplate} />
-                    <Column field="activity" header="Activity" showFilterMatchModes={false} style={{ minWidth: '12rem' }} body={activityBodyTemplate} filter filterElement={activityFilterTemplate} />
-                    <Column field="verified" header="Verified" dataType="boolean" bodyClassName="text-center" style={{ minWidth: '8rem' }} body={verifiedBodyTemplate} filter filterElement={verifiedFilterTemplate} />
-                </DataTable>
-            </div>
-        </div>
-    );
-}
-        `,
-        extFiles: {
-            'DataTableDemo.css': `
-/* DataTableDemo.css */
-
-.datatable-filter-demo .p-paginator .p-paginator-current {
-    margin-left: auto;
-}
-.datatable-filter-demo .p-progressbar {
-    height: 0.5rem;
-    background-color: #d8dadc;
-}
-.datatable-filter-demo .p-progressbar .p-progressbar-value {
-    background-color: #607d8b;
-}
-.datatable-filter-demo .p-datepicker {
-    min-width: 25rem;
-}
-.datatable-filter-demo .p-datepicker td {
-    font-weight: 400;
-}
-.datatable-filter-demo .p-datatable.p-datatable-customers .p-datatable-header {
-    padding: 1rem;
-    text-align: left;
-    font-size: 1.5rem;
-}
-.datatable-filter-demo .p-datatable.p-datatable-customers .p-paginator {
-    padding: 1rem;
-}
-.datatable-filter-demo .p-datatable.p-datatable-customers .p-datatable-thead > tr > th {
-    text-align: left;
-}
-.datatable-filter-demo .p-datatable.p-datatable-customers .p-datatable-tbody > tr > td {
-    cursor: auto;
-}
-.datatable-filter-demo .p-datatable.p-datatable-customers .p-dropdown-label:not(.p-placeholder) {
-    text-transform: uppercase;
-}
-            `
-        },
-        data: `
-/* CustomerService */ 
-{
-    id: 1000,
-    name: 'James Butt',
-    country: {
-        name: 'Algeria',
-        code: 'dz'
-    },
-    company: 'Benton, John B Jr',
-    date: '2015-09-13',
-    status: 'unqualified',
-    verified: true,
-    activity: 17,
-    representative: {
-        name: 'Ioni Bowcher',
-        image: 'ionibowcher.png'
-    },
-    balance: 70663
-},
-...
-       `
-    };
-
-    return (
-        <>
-            <DocSectionText {...props}>
-                <p>Filters are displayed in an overlay.</p>{' '}
-            </DocSectionText>
-            <div className="card datatable-filter-demo ">
-                <DataTable
-                    value={customers}
-                    paginator
-                    className="p-datatable-customers"
-                    showGridlines
-                    rows={10}
-                    dataKey="id"
-                    filters={filters}
-                    filterDisplay="menu"
-                    loading={loading}
-                    responsiveLayout="scroll"
-                    globalFilterFields={['name', 'country.name', 'representative.name', 'balance', 'status']}
-                    header={header}
-                    emptyMessage="No customers found."
-                >
-                    <Column field="name" header="Name" filter filterPlaceholder="Search by name" style={{ minWidth: '12rem' }} />
-                    <Column
-                        header="Country"
-                        filterField="country.name"
-                        style={{ minWidth: '12rem' }}
-                        body={countryBodyTemplate}
-                        filter
-                        filterPlaceholder="Search by country"
-                        filterClear={filterClearTemplate}
-                        filterApply={filterApplyTemplate}
-                        filterFooter={filterFooterTemplate}
-                    />
-                    <Column
-                        header="Agent"
-                        filterField="representative"
-                        showFilterMatchModes={false}
-                        filterMenuStyle={{ width: '14rem' }}
-                        style={{ minWidth: '14rem' }}
-                        body={representativeBodyTemplate}
-                        filter
-                        filterElement={representativeFilterTemplate}
-                    />
-                    <Column header="Date" filterField="date" dataType="date" style={{ minWidth: '10rem' }} body={dateBodyTemplate} filter filterElement={dateFilterTemplate} />
-                    <Column header="Balance" filterField="balance" dataType="numeric" style={{ minWidth: '10rem' }} body={balanceBodyTemplate} filter filterElement={balanceFilterTemplate} />
-                    <Column field="status" header="Status" filterMenuStyle={{ width: '14rem' }} style={{ minWidth: '12rem' }} body={statusBodyTemplate} filter filterElement={statusFilterTemplate} />
-                    <Column field="activity" header="Activity" showFilterMatchModes={false} style={{ minWidth: '12rem' }} body={activityBodyTemplate} filter filterElement={activityFilterTemplate} />
-                    <Column field="verified" header="Verified" dataType="boolean" bodyClassName="text-center" style={{ minWidth: '8rem' }} body={verifiedBodyTemplate} filter filterElement={verifiedFilterTemplate} />
-                </DataTable>
-            </div>
-            <DocSectionCode code={code} service={['CustomerService']} />
-        </>
-    );
-}
