+++ service/CustomerService.js
@@ -28,7 +28,7 @@ export const CustomerService = {
                 },
                 company: 'Chanay, Jeffrey A Esq',
                 date: '2019-02-09',
-                status: 'proposal',
+                status: 'negotiation',
                 verified: true,
                 activity: 0,
                 representative: {
@@ -82,7 +82,7 @@ export const CustomerService = {
                 },
                 company: 'Printing Dimensions',
                 date: '2016-05-20',
-                status: 'proposal',
+                status: 'negotiation',
                 verified: true,
                 activity: 33,
                 representative: {
@@ -172,7 +172,7 @@ export const CustomerService = {
                 },
                 company: 'King, Christopher A Esq',
                 date: '2015-07-07',
-                status: 'proposal',
+                status: 'negotiation',
                 verified: false,
                 activity: 3,
                 representative: {
@@ -604,7 +604,7 @@ export const CustomerService = {
                 },
                 company: 'Giampetro, Anthony D',
                 date: '2019-08-08',
-                status: 'proposal',
+                status: 'negotiation',
                 verified: true,
                 activity: 85,
                 representative: {
@@ -1108,7 +1108,7 @@ export const CustomerService = {
                 },
                 company: 'Hubbard, Bruce Esq',
                 date: '2017-05-25',
-                status: 'proposal',
+                status: 'negotiation',
                 verified: true,
                 activity: 87,
                 representative: {
@@ -1342,7 +1342,7 @@ export const CustomerService = {
                 },
                 company: 'A K Construction Co',
                 date: '2018-11-13',
-                status: 'proposal',
+                status: 'negotiation',
                 verified: true,
                 activity: 22,
                 representative: {
@@ -1432,7 +1432,7 @@ export const CustomerService = {
                 },
                 company: 'Museum Of Science & Industry',
                 date: '2017-06-01',
-                status: 'proposal',
+                status: 'negotiation',
                 verified: true,
                 activity: 90,
                 representative: {
@@ -1450,7 +1450,7 @@ export const CustomerService = {
                 },
                 company: 'Garrison Ind',
                 date: '2017-03-14',
-                status: 'proposal',
+                status: 'negotiation',
                 verified: true,
                 activity: 96,
                 representative: {
@@ -1486,7 +1486,7 @@ export const CustomerService = {
                 },
                 company: 'Smc Inc',
                 date: '2018-09-11',
-                status: 'proposal',
+                status: 'negotiation',
                 verified: true,
                 activity: 46,
                 representative: {
@@ -1702,7 +1702,7 @@ export const CustomerService = {
                 },
                 company: 'Tree Musketeers',
                 date: '2019-02-17',
-                status: 'proposal',
+                status: 'negotiation',
                 verified: true,
                 activity: 50,
                 representative: {
@@ -1756,7 +1756,7 @@ export const CustomerService = {
                 },
                 company: 'Kelly, Charles G Esq',
                 date: '2019-10-02',
-                status: 'proposal',
+                status: 'negotiation',
                 verified: true,
                 activity: 44,
                 representative: {
@@ -1792,7 +1792,7 @@ export const CustomerService = {
                 },
                 company: 'Beck Horizon Builders',
                 date: '2018-05-14',
-                status: 'proposal',
+                status: 'negotiation',
                 verified: true,
                 activity: 99,
                 representative: {
@@ -1900,7 +1900,7 @@ export const CustomerService = {
                 },
                 company: 'American Arts & Graphics',
                 date: '2020-06-18',
-                status: 'proposal',
+                status: 'negotiation',
                 verified: true,
                 activity: 21,
                 representative: {
@@ -2008,7 +2008,7 @@ export const CustomerService = {
                 },
                 company: 'Juno Chefs Incorporated',
                 date: '2017-03-17',
-                status: 'proposal',
+                status: 'negotiation',
                 verified: false,
                 activity: 31,
                 representative: {
@@ -2116,7 +2116,7 @@ export const CustomerService = {
                 },
                 company: 'Jackson Shields Yeiser',
                 date: '2016-07-13',
-                status: 'proposal',
+                status: 'negotiation',
                 verified: true,
                 activity: 55,
                 representative: {
@@ -2188,7 +2188,7 @@ export const CustomerService = {
                 },
                 company: 'Carol, Drake Sparks Esq',
                 date: '2018-05-22',
-                status: 'proposal',
+                status: 'negotiation',
                 verified: true,
                 activity: 77,
                 representative: {
@@ -2242,7 +2242,7 @@ export const CustomerService = {
                 },
                 company: 'Merlin Electric Co',
                 date: '2019-05-23',
-                status: 'proposal',
+                status: 'negotiation',
                 verified: true,
                 activity: 10,
                 representative: {
@@ -2332,7 +2332,7 @@ export const CustomerService = {
                 },
                 company: 'Nelson Hawaiian Ltd',
                 date: '2016-11-10',
-                status: 'proposal',
+                status: 'negotiation',
                 verified: true,
                 activity: 41,
                 representative: {
@@ -2476,7 +2476,7 @@ export const CustomerService = {
                 },
                 company: 'Vlahos, John J Esq',
                 date: '2017-03-12',
-                status: 'proposal',
+                status: 'negotiation',
                 verified: false,
                 activity: 3,
                 representative: {
@@ -2566,7 +2566,7 @@ export const CustomerService = {
                 },
                 company: 'Admiral Party Rentals & Sales',
                 date: '2016-02-10',
-                status: 'proposal',
+                status: 'negotiation',
                 verified: false,
                 activity: 58,
                 representative: {
@@ -2638,7 +2638,7 @@ export const CustomerService = {
                 },
                 company: 'Megibow & Edwards',
                 date: '2019-08-17',
-                status: 'proposal',
+                status: 'negotiation',
                 verified: true,
                 activity: 18,
                 representative: {
@@ -2764,7 +2764,7 @@ export const CustomerService = {
                 },
                 company: 'Silver Bros Inc',
                 date: '2016-08-19',
-                status: 'proposal',
+                status: 'negotiation',
                 verified: true,
                 activity: 14,
                 representative: {
@@ -2782,7 +2782,7 @@ export const CustomerService = {
                 },
                 company: 'Newtec Inc',
                 date: '2020-02-09',
-                status: 'proposal',
+                status: 'negotiation',
                 verified: true,
                 activity: 77,
                 representative: {
@@ -3196,7 +3196,7 @@ export const CustomerService = {
                 },
                 company: 'Corn Popper',
                 date: '2020-09-16',
-                status: 'proposal',
+                status: 'negotiation',
                 verified: true,
                 activity: 76,
                 representative: {
@@ -3268,7 +3268,7 @@ export const CustomerService = {
                 },
                 company: 'Royal Pontiac Olds Inc',
                 date: '2018-07-19',
-                status: 'proposal',
+                status: 'negotiation',
                 verified: false,
                 activity: 24,
                 representative: {
@@ -3286,7 +3286,7 @@ export const CustomerService = {
                 },
                 company: 'Leonards Antiques Inc',
                 date: '2015-03-15',
-                status: 'proposal',
+                status: 'negotiation',
                 verified: true,
                 activity: 86,
                 representative: {
@@ -3322,7 +3322,7 @@ export const CustomerService = {
                 },
                 company: 'Nercon Engineering & Mfg Inc',
                 date: '2015-02-19',
-                status: 'proposal',
+                status: 'negotiation',
                 verified: true,
                 activity: 9,
                 representative: {
@@ -3448,7 +3448,7 @@ export const CustomerService = {
                 },
                 company: 'Computer Repair Service',
                 date: '2016-02-13',
-                status: 'proposal',
+                status: 'negotiation',
                 verified: false,
                 activity: 79,
                 representative: {
@@ -3502,7 +3502,7 @@ export const CustomerService = {
                 },
                 company: 'New England Taxidermy',
                 date: '2019-01-07',
-                status: 'proposal',
+                status: 'negotiation',
                 verified: true,
                 activity: 4,
                 representative: {
@@ -3880,7 +3880,7 @@ export const CustomerService = {
                 },
                 company: 'Metlab Testing Services',
                 date: '2015-02-13',
-                status: 'proposal',
+                status: 'negotiation',
                 verified: true,
                 activity: 99,
                 representative: {
@@ -3970,7 +3970,7 @@ export const CustomerService = {
                 },
                 company: 'Dill Dill Carr \u0026 Stonbraker Pc',
                 date: '2017-07-27',
-                status: 'proposal',
+                status: 'negotiation',
                 verified: true,
                 activity: 63,
                 representative: {
@@ -4096,7 +4096,7 @@ export const CustomerService = {
                 },
                 company: 'Little Sheet Metal Co',
                 date: '2016-06-22',
-                status: 'proposal',
+                status: 'negotiation',
                 verified: true,
                 activity: 38,
                 representative: {
@@ -4384,7 +4384,7 @@ export const CustomerService = {
                 },
                 company: 'Bosco, Paul J',
                 date: '2017-09-11',
-                status: 'proposal',
+                status: 'negotiation',
                 verified: true,
                 activity: 60,
                 representative: {
@@ -4474,7 +4474,7 @@ export const CustomerService = {
                 },
                 company: 'Linguistic Systems Inc',
                 date: '2017-08-04',
-                status: 'proposal',
+                status: 'negotiation',
                 verified: true,
                 activity: 86,
                 representative: {
@@ -4726,7 +4726,7 @@ export const CustomerService = {
                 },
                 company: 'Price Business Services',
                 date: '2019-10-11',
-                status: 'proposal',
+                status: 'negotiation',
                 verified: true,
                 activity: 98,
                 representative: {
@@ -4978,7 +4978,7 @@ export const CustomerService = {
                 },
                 company: 'Schott Fiber Optics Inc',
                 date: '2019-06-05',
-                status: 'proposal',
+                status: 'negotiation',
                 verified: true,
                 activity: 95,
                 representative: {
@@ -5140,7 +5140,7 @@ export const CustomerService = {
                 },
                 company: 'Geonex Martel Inc',
                 date: '2015-03-19',
-                status: 'proposal',
+                status: 'negotiation',
                 verified: true,
                 activity: 64,
                 representative: {
@@ -5248,7 +5248,7 @@ export const CustomerService = {
                 },
                 company: 'Flash, Elena Salerno Esq',
                 date: '2018-01-28',
-                status: 'proposal',
+                status: 'negotiation',
                 verified: true,
                 activity: 0,
                 representative: {
@@ -5284,7 +5284,7 @@ export const CustomerService = {
                 },
                 company: 'Alabama Educational Tv Comm',
                 date: '2019-06-09',
-                status: 'proposal',
+                status: 'negotiation',
                 verified: true,
                 activity: 30,
                 representative: {
@@ -5302,7 +5302,7 @@ export const CustomerService = {
                 },
                 company: 'Franz Inc',
                 date: '2017-02-07',
-                status: 'proposal',
+                status: 'negotiation',
                 verified: true,
                 activity: 59,
                 representative: {
@@ -5392,7 +5392,7 @@ export const CustomerService = {
                 },
                 company: 'Grayson, Grant S Esq',
                 date: '2019-12-22',
-                status: 'proposal',
+                status: 'negotiation',
                 verified: true,
                 activity: 22,
                 representative: {
@@ -5446,7 +5446,7 @@ export const CustomerService = {
                 },
                 company: 'Wmmt Radio Station',
                 date: '2016-03-28',
-                status: 'proposal',
+                status: 'negotiation',
                 verified: true,
                 activity: 73,
                 representative: {
@@ -5536,7 +5536,7 @@ export const CustomerService = {
                 },
                 company: 'Kaminski, Katherine Andritsaki',
                 date: '2015-07-17',
-                status: 'proposal',
+                status: 'negotiation',
                 verified: true,
                 activity: 29,
                 representative: {
@@ -5608,7 +5608,7 @@ export const CustomerService = {
                 },
                 company: 'Anchor Computer Inc',
                 date: '2018-12-08',
-                status: 'proposal',
+                status: 'negotiation',
                 verified: true,
                 activity: 89,
                 representative: {
@@ -5644,7 +5644,7 @@ export const CustomerService = {
                 },
                 company: 'H A C Farm Lines Co Optv Assoc',
                 date: '2017-02-17',
-                status: 'proposal',
+                status: 'negotiation',
                 verified: true,
                 activity: 58,
                 representative: {
@@ -5734,7 +5734,7 @@ export const CustomerService = {
                 },
                 company: 'Schoen, Edward J Jr',
                 date: '2017-03-03',
-                status: 'proposal',
+                status: 'negotiation',
                 verified: true,
                 activity: 29,
                 representative: {
@@ -5932,7 +5932,7 @@ export const CustomerService = {
                 },
                 company: 'National Medical Excess Corp',
                 date: '2019-09-01',
-                status: 'proposal',
+                status: 'negotiation',
                 verified: true,
                 activity: 8,
                 representative: {
@@ -5986,7 +5986,7 @@ export const CustomerService = {
                 },
                 company: 'Insty Prints',
                 date: '2019-02-02',
-                status: 'proposal',
+                status: 'negotiation',
                 verified: true,
                 activity: 8,
                 representative: {
@@ -6022,7 +6022,7 @@ export const CustomerService = {
                 },
                 company: 'Circuit Solution Inc',
                 date: '2019-07-10',
-                status: 'proposal',
+                status: 'negotiation',
                 verified: true,
                 activity: 88,
                 representative: {
@@ -6040,7 +6040,7 @@ export const CustomerService = {
                 },
                 company: 'Local Office',
                 date: '2016-05-21',
-                status: 'proposal',
+                status: 'negotiation',
                 verified: true,
                 activity: 85,
                 representative: {
@@ -6076,7 +6076,7 @@ export const CustomerService = {
                 },
                 company: 'Davis, Maxon R Esq',
                 date: '2019-06-20',
-                status: 'proposal',
+                status: 'negotiation',
                 verified: true,
                 activity: 44,
                 representative: {
@@ -6094,7 +6094,7 @@ export const CustomerService = {
                 },
                 company: 'Garrison, Paul E Esq',
                 date: '2019-08-01',
-                status: 'proposal',
+                status: 'negotiation',
                 verified: true,
                 activity: 91,
                 representative: {
@@ -6220,7 +6220,7 @@ export const CustomerService = {
                 },
                 company: 'H P G Industrys Inc',
                 date: '2019-12-15',
-                status: 'proposal',
+                status: 'negotiation',
                 verified: true,
                 activity: 20,
                 representative: {
@@ -6310,7 +6310,7 @@ export const CustomerService = {
                 },
                 company: 'Logan, Ronald J Esq',
                 date: '2016-02-13',
-                status: 'proposal',
+                status: 'negotiation',
                 verified: true,
                 activity: 5,
                 representative: {
@@ -6346,7 +6346,7 @@ export const CustomerService = {
                 },
                 company: 'Germer And Gertz Llp',
                 date: '2015-01-03',
-                status: 'proposal',
+                status: 'negotiation',
                 verified: true,
                 activity: 66,
                 representative: {
@@ -6454,7 +6454,7 @@ export const CustomerService = {
                 },
                 company: 'Quality Inn Northwest',
                 date: '2016-04-05',
-                status: 'proposal',
+                status: 'negotiation',
                 verified: true,
                 activity: 14,
                 representative: {
@@ -6598,7 +6598,7 @@ export const CustomerService = {
                 },
                 company: 'Gehring Pumps Inc',
                 date: '2018-10-21',
-                status: 'proposal',
+                status: 'negotiation',
                 verified: true,
                 activity: 90,
                 representative: {
@@ -6706,7 +6706,7 @@ export const CustomerService = {
                 },
                 company: 'Minor, Cynthia A Esq',
                 date: '2020-09-25',
-                status: 'proposal',
+                status: 'negotiation',
                 verified: true,
                 activity: 17,
                 representative: {
@@ -6850,7 +6850,7 @@ export const CustomerService = {
                 },
                 company: 'Soward, Anne Esq',
                 date: '2018-10-07',
-                status: 'proposal',
+                status: 'negotiation',
                 verified: false,
                 activity: 28,
                 representative: {
@@ -6976,7 +6976,7 @@ export const CustomerService = {
                 },
                 company: 'Sameshima, Douglas J Esq',
                 date: '2019-11-27',
-                status: 'proposal',
+                status: 'negotiation',
                 verified: true,
                 activity: 85,
                 representative: {
@@ -6994,7 +6994,7 @@ export const CustomerService = {
                 },
                 company: 'Thompson, John Randolph Jr',
                 date: '2017-08-13',
-                status: 'proposal',
+                status: 'negotiation',
                 verified: true,
                 activity: 40,
                 representative: {
@@ -7120,7 +7120,7 @@ export const CustomerService = {
                 },
                 company: 'Vico Products Mfg Co',
                 date: '2017-10-02',
-                status: 'proposal',
+                status: 'negotiation',
                 verified: true,
                 activity: 53,
                 representative: {
@@ -7156,7 +7156,7 @@ export const CustomerService = {
                 },
                 company: 'Farrell \u0026 Johnson Office Equip',
                 date: '2017-06-24',
-                status: 'proposal',
+                status: 'negotiation',
                 verified: true,
                 activity: 39,
                 representative: {
@@ -7264,7 +7264,7 @@ export const CustomerService = {
                 },
                 company: 'Barkan, Neal J Esq',
                 date: '2019-02-19',
-                status: 'proposal',
+                status: 'negotiation',
                 verified: true,
                 activity: 38,
                 representative: {
@@ -7318,7 +7318,7 @@ export const CustomerService = {
                 },
                 company: 'Young, Timothy L Esq',
                 date: '2019-02-05',
-                status: 'proposal',
+                status: 'negotiation',
                 verified: true,
                 activity: 55,
                 representative: {
@@ -7408,7 +7408,7 @@ export const CustomerService = {
                 },
                 company: 'Sat Poly Inc',
                 date: '2020-10-19',
-                status: 'proposal',
+                status: 'negotiation',
                 verified: true,
                 activity: 69,
                 representative: {
@@ -7516,7 +7516,7 @@ export const CustomerService = {
                 },
                 company: 'United Cerebral Palsy Of Ne Pa',
                 date: '2020-12-11',
-                status: 'proposal',
+                status: 'negotiation',
                 verified: true,
                 activity: 0,
                 representative: {
@@ -7732,7 +7732,7 @@ export const CustomerService = {
                 },
                 company: 'Armon Communications',
                 date: '2018-02-16',
-                status: 'proposal',
+                status: 'negotiation',
                 verified: true,
                 activity: 84,
                 representative: {
@@ -7786,7 +7786,7 @@ export const CustomerService = {
                 },
                 company: 'Haughey, Charles Jr',
                 date: '2016-11-03',
-                status: 'proposal',
+                status: 'negotiation',
                 verified: true,
                 activity: 63,
                 representative: {
@@ -7840,7 +7840,7 @@ export const CustomerService = {
                 },
                 company: 'Nazette Marner Good Wendt',
                 date: '2015-08-18',
-                status: 'proposal',
+                status: 'negotiation',
                 verified: true,
                 activity: 66,
                 representative: {
@@ -7858,7 +7858,7 @@ export const CustomerService = {
                 },
                 company: 'Levinson Axelrod Wheaton',
                 date: '2016-02-11',
-                status: 'proposal',
+                status: 'negotiation',
                 verified: true,
                 activity: 64,
                 representative: {
@@ -8128,7 +8128,7 @@ export const CustomerService = {
                 },
                 company: 'Dimmock, Thomas J Esq',
                 date: '2017-02-10',
-                status: 'proposal',
+                status: 'negotiation',
                 verified: true,
                 activity: 9,
                 representative: {
@@ -8272,7 +8272,7 @@ export const CustomerService = {
                 },
                 company: 'Switchcraft Inc',
                 date: '2019-06-08',
-                status: 'proposal',
+                status: 'negotiation',
                 verified: true,
                 activity: 21,
                 representative: {
@@ -8308,7 +8308,7 @@ export const CustomerService = {
                 },
                 company: 'Shimotani, Grace T',
                 date: '2020-11-15',
-                status: 'proposal',
+                status: 'negotiation',
                 verified: true,
                 activity: 94,
                 representative: {
@@ -8560,7 +8560,7 @@ export const CustomerService = {
                 },
                 company: 'Mcrae, James L',
                 date: '2018-03-20',
-                status: 'proposal',
+                status: 'negotiation',
                 verified: true,
                 activity: 45,
                 representative: {
@@ -8578,7 +8578,7 @@ export const CustomerService = {
                 },
                 company: 'Mcgraths Seafood',
                 date: '2017-05-13',
-                status: 'proposal',
+                status: 'negotiation',
                 verified: true,
                 activity: 34,
                 representative: {
@@ -8596,7 +8596,7 @@ export const CustomerService = {
                 },
                 company: 'Kummerer, J Michael Esq',
                 date: '2017-02-21',
-                status: 'proposal',
+                status: 'negotiation',
                 verified: true,
                 activity: 12,
                 representative: {
